refactor: decompose HumanoidLauncher into UISetupMixin, SimulationMixin, AnalysisMixin
Extract methods into 3 mixin files:
- humanoid_launcher_ui.py (555 lines): Tab/widget setup for all 4 tabs + log area
- humanoid_launcher_sim.py (240 lines): Docker commands, simulation launch/stop, file opening
- humanoid_launcher_analysis.py (461 lines): IAA plotting, config save/load, dialogs, logging

Reduces humanoid_launcher.py from 1,584 to 268 lines (-83.1%)
